Combining different implementations of the BitMenuButton component in just one component #3921
