+++ src/BlazorUI/Bit.BlazorUI/Components/Buttons/BitMenuButton/BitMenuButtonJsExtensions.cs
@@ -3,13 +3,13 @@ namespace Bit.BlazorUI;
 
 internal static class BitMenuButtonJsExtensions
 {
-    internal static async Task ToggleMenuButtonCallout(this IJSRuntime jsRuntime,
-                                                           DotNetObjectReference<BitMenuButton> dotNetObjRef,
-                                                           string uniqueId, 
-                                                           string menuButtonId,
-                                                           string menuButtonCalloutId, 
-                                                           string menuButtonOverlayId,
-                                                           bool isCalloutOpen)
+    internal static async Task ToggleMenuButtonCallout<T>(this IJSRuntime jsRuntime,
+                                                              DotNetObjectReference<BitMenuButton<T>> dotNetObjRef,
+                                                              string uniqueId,
+                                                              string menuButtonId,
+                                                              string menuButtonCalloutId,
+                                                              string menuButtonOverlayId,
+                                                              bool isCalloutOpen) where T : class
     {
         await jsRuntime.InvokeVoidAsync("BitMenuButton.toggleMenuButtonCallout", dotNetObjRef, uniqueId, menuButtonId, menuButtonCalloutId, menuButtonOverlayId, isCalloutOpen);
     }
