--- src/BlazorUI/Bit.BlazorUI/Components/Buttons/BitMenuButtonGroup/BitMenuButtonGroup.razor
@@ -1,41 +1,0 @@
-﻿@namespace Bit.BlazorUI
-@inherits BitComponentBase
-
-<div @ref="RootElement"
-     @attributes="HtmlAttributes"
-     style="@StyleBuilder.Value"
-     class="@ClassBuilder.Value"
-     id="@UniqueId">
-    <div id="@_menuButtonOverlayId" class="overlay" @onclick="CloseCallout"></div>
-    <button id="@_menuButtonId"
-            type="@ButtonType?.GetValue()"
-            aria-label="@AriaLabel"
-            aria-describedby="@AriaDescription"
-            aria-hidden="@AriaHidden"
-            tabindex="@(IsEnabled ? 0 : -1)"
-            class="menu-btn"
-            @onclick="HandleOnClick">
-        @if (HeaderTemplate is not null)
-        {
-            @HeaderTemplate
-        }
-        else
-        {
-            @if (IconName.HasValue)
-            {
-                <i class="bit-icon bit-icon--@IconName.GetName()"></i>
-            }
-            <span class="text">@Text</span>
-            <i class="bit-icon bit-icon--ChevronDown"></i>
-        }
-    </button>
-    <div id="@_menuButtonCalloutId"
-         class="options"
-         tabindex="@(IsEnabled ? 0 : -1)">
-        <ul role="presentation">
-            <CascadingValue Value="this" IsFixed="true">
-                @ChildContent
-            </CascadingValue>
-        </ul>
-    </div>
-</div>
