--- src/BlazorUI/Bit.BlazorUI/Components/Buttons/BitMenuButtonGroup/BitMenuButtonGroup.scss
@@ -1,180 +1,0 @@
-﻿@import "../../../Styles/functions.scss";
-
-.bit-mbg {
-    position: relative;
-    min-width: spacing(10);
-    height: spacing(4);
-    font-family: $typography-button-font-family;
-    font-size: spacing(1.75);
-    font-weight: 400;
-    display: inline-flex;
-    justify-content: center;
-    align-items: center;
-    outline: transparent;
-    -webkit-font-smoothing: antialiased;
-    border-radius: $shape-border-radius;
-    border-width: $shape-border-width;
-    border-style: $shape-border-style;
-    cursor: pointer;
-
-    .text {
-        margin: 0 spacing(0.5);
-    }
-
-    .bit-icon {
-        margin: spacing(0.25) spacing(0.5) 0 spacing(0.5);
-    }
-
-    &.disabled {
-        color: $color-text-disabled;
-        background-color: $color-action-disabled;
-        border-color: $color-border-disabled;
-        cursor: default;
-        pointer-events: none;
-    }
-
-    &.standard {
-        border-color: $color-border-secondary;
-        color: $color-text-secondary;
-        background-color: $color-secondary-main;
-
-        &:hover {
-            border-color: $color-action-borde-hover-secondary;
-            background-color: $color-action-hover-secondary;
-        }
-
-        &:active, &.open-menu {
-            border-color: $color-action-borde-active-secondary;
-            background-color: $color-action-active-secondary;
-        }
-    }
-
-    &.primary {
-        border-color: $color-border-primary;
-        color: $color-text-primary;
-        background-color: $color-primary-main;
-
-        &:hover {
-            border-color: $color-action-hover-primary;
-            background-color: $color-action-hover-primary;
-        }
-
-        &:active, &.open-menu {
-            border-color: $color-action-active-primary;
-            background-color: $color-action-active-primary;
-        }
-    }
-
-    &.open-menu {
-        .options {
-            opacity: 1;
-            pointer-events: auto;
-            transform: translate3d(0, 0, 0);
-        }
-    }
-
-    .menu-btn {
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        background-color: transparent;
-        width: 100%;
-        height: 100%;
-        cursor: inherit;
-        pointer-events: inherit;
-        color: inherit;
-        padding: 0 spacing(2);
-
-        .text {
-            margin: 0 spacing(0.5);
-        }
-
-        .bit-icon {
-            margin: spacing(0.25) spacing(0.5) 0 spacing(0.5);
-        }
-    }
-
-    .overlay {
-        display: none;
-        width: 100%;
-        height: 100%;
-        position: fixed;
-        top: 0;
-        left: 0;
-        z-index: 999;
-        background-color: transparent;
-    }
-
-    .options {
-        display: none;
-        position: fixed;
-        min-width: spacing(22.5);
-        background-color: $color-neutrals-white;
-        box-sizing: border-box;
-        border-radius: 0 0 spacing(0.25) spacing(0.25);
-        box-shadow: rgba(0, 0, 0, 0.13) 0 spacing(0.4) spacing(0.9) 0, rgba(0, 0, 0, 0.11) 0 spacing(0.075) spacing(0.225) 0;
-        z-index: 1001;
-        overflow: hidden auto;
-        -webkit-font-smoothing: antialiased;
-        opacity: 0;
-        transform: translate3d(0, spacing(-1.25), 0);
-        pointer-events: none;
-        transition: opacity 100ms, transform 300ms, pointer-events 300ms;
-
-        ul {
-            list-style-type: none;
-            margin: 0;
-            padding: 0;
-            width: 100%;
-        }
-    }
-
-    .option {
-        display: flex;
-        background-color: transparent;
-        width: 100%;
-        height: spacing(4.5);
-        line-height: spacing(4.5);
-        align-items: center;
-        text-align: left;
-        cursor: pointer;
-        padding: 0 spacing(1) 0 spacing(0.5);
-        border: none;
-        color: $color-text-secondary;
-
-        .bit-icon {
-            color: $color-primary-main;
-            margin: 0 spacing(0.5);
-        }
-
-        .text {
-            margin: 0 spacing(0.5);
-            vertical-align: middle;
-            display: inline-block;
-            flex-grow: 1;
-            text-overflow: ellipsis;
-            white-space: nowrap;
-        }
-
-        &.disabled {
-            cursor: default;
-            color: $color-text-disabled;
-
-            .bit-icon {
-                color: $color-text-disabled;
-            }
-
-            &:hover {
-                background-color: transparent;
-            }
-        }
-
-        &:hover {
-            background-color: $color-action-hover-secondary;
-        }
-
-        &:active {
-            background-color: $color-action-active-secondary;
-        }
-    }
-}
