--- src/BlazorUI/Bit.BlazorUI/Components/Buttons/BitMenuButtonList/BitMenuButtonList.razor
@@ -1,65 +1,0 @@
-﻿@namespace Bit.BlazorUI
-@inherits BitComponentBase
-@typeparam TItem
-
-<div @ref="RootElement"
-     @attributes="HtmlAttributes"
-     style="@StyleBuilder.Value"
-     class="@ClassBuilder.Value"
-     id="@UniqueId">
-    <div id="@_menuButtonOverlayId" class="overlay" @onclick="CloseCallout"></div>
-    <button id="@_menuButtonId"
-            type="@ButtonType?.GetValue()"
-            aria-label="@AriaLabel"
-            aria-describedby="@AriaDescription"
-            aria-hidden="@AriaHidden"
-            tabindex="@(IsEnabled ? 0 : -1)"
-            class="menu-btn"
-            @onclick="HandleOnClick">
-        @if (HeaderTemplate is not null)
-        {
-            @HeaderTemplate
-        }
-        else
-        {
-            @if (IconName.HasValue)
-            {
-                <i class="bit-icon bit-icon--@IconName.GetName()"></i>
-            }
-            <span class="text">@Text</span>
-            <i class="bit-icon bit-icon--ChevronDown"></i>
-        }
-    </button>
-    <div id="@_menuButtonCalloutId"
-         class="items"
-         tabindex="@(IsEnabled ? 0 : -1)">
-        <ul role="presentation">
-            @foreach (TItem item in Items)
-            {
-                <li role="presentation">
-                    <button @onclick="() => HandleOnItemClick(item)"
-                            type="@ButtonType?.GetValue()"
-                            role="menuitem"
-                            aria-disabled="@GetIsEnabled(item)"
-                            tabindex="@(GetIsEnabled(item) ? 0 : -1)"
-                            class="item @(GetIsEnabled(item) ? "" : "disabled")">
-                        @if (ItemTemplate is not null)
-                        {
-                            @ItemTemplate(item)
-                        }
-                        else
-                        {
-                            @if (GetIconName(item).HasValue)
-                            {
-                                <i class="bit-icon bit-icon--@GetIconName(item).GetName()"></i>
-                            }
-                            <span class="text">
-                                @GetText(item)
-                            </span>
-                        }
-                    </button>
-                </li>
-            }
-        </ul>
-    </div>
-</div>
