+++ src/BlazorUI/Bit.BlazorUI/Styles/Fluent/bit.blazorui.fluent.scss
@@ -19,8 +19,6 @@
 @import "../../Components/Buttons/BitSplitButton/BitSplitButton.scss";
 @import "../../Components/Buttons/BitSplitButtonList/BitSplitButtonList.scss";
 @import "../../Components/Buttons/BitMenuButton/BitMenuButton.scss";
-@import "../../Components/Buttons/BitMenuButtonGroup/BitMenuButtonGroup.scss";
-@import "../../Components/Buttons/BitMenuButtonList/BitMenuButtonList.scss";
 @import "../../Components/Buttons/BitToggleButton/BitToggleButton.scss";
 @import "../../Components/Carousel/BitCarousel.Fluent.scss";
 @import "../../Components/Carousel/BitCarouselItem.Fluent.scss";
