+++ src/BlazorUI/Playground/Bit.BlazorUI.Playground/Web/Pages/Components/Buttons/BitMenuButtonDemo.razor.scss
@@ -1,14 +1,9 @@
 ﻿.example-content {
     display: flex;
     flex-flow: row wrap;
-    gap: 0.5rem;
-    width: fit-content;
     align-items: center;
-
-    &.column {
-        flex-direction: column;
-        align-items: flex-start;
-    }
+    gap: 0.5rem;
+    margin-bottom: 30px;
 }
 
 .selected-item {
