--- src/BlazorUI/Playground/Bit.BlazorUI.Playground/Web/Pages/Components/Buttons/BitMenuButtonGroupDemo.razor.scss
@@ -1,36 +1,0 @@
-﻿.example-content {
-    display: flex;
-    flex-flow: row wrap;
-    gap: 0.5rem;
-    width: fit-content;
-    align-items: center;
-
-    &.column {
-        flex-direction: column;
-        align-items: flex-start;
-    }
-}
-
-.selected-item {
-    margin-top: 0.5rem;
-    font-weight: 600;
-}
-
-::deep .custom-menu-btn {
-    &.primary {
-        height: 2.5rem;
-        width: 10.5rem;
-        background-color: #515151;
-        border-color: black;
-
-        &:hover {
-            background-color: #403f3f;
-            border-color: black;
-        }
-    }
-}
-
-.item-template-box {
-    display: flex;
-    width: 100%;
-}
