+++ src/BlazorUI/Playground/Bit.BlazorUI.Playground/Web/Pages/HomePage.razor
@@ -170,12 +170,6 @@
                         <BitLink Class="component-list-item" Href="/components/menu-button" title="MenuButton">
                             <span>MenuButton</span>
                         </BitLink>
-                        <BitLink Class="component-list-item" Href="/components/menu-button-group" title="MenuButtonGroup">
-                            <span>MenuButtonGroup</span>
-                        </BitLink>
-                        <BitLink Class="component-list-item" Href="/components/menu-button-list" title="MenuButtonList">
-                            <span>MenuButtonList</span>
-                        </BitLink>
                         <BitLink Class="component-list-item" Href="/components/split-button" title="SplitButton">
                             <span>SplitButton</span>
                         </BitLink>
