+++ src/BlazorUI/Playground/Bit.BlazorUI.Playground/Web/compilerconfig.json
@@ -72,14 +72,6 @@
         "inputFile": "Pages/Components/Buttons/BitMenuButtonDemo.razor.scss"
     },
     {
-        "outputFile": "Pages/Components/Buttons/BitMenuButtonGroupDemo.razor.css",
-        "inputFile": "Pages/Components/Buttons/BitMenuButtonGroupDemo.razor.scss"
-    },
-    {
-        "outputFile": "Pages/Components/Buttons/BitMenuButtonListDemo.razor.css",
-        "inputFile": "Pages/Components/Buttons/BitMenuButtonListDemo.razor.scss"
-    },
-    {
         "outputFile": "Pages/Components/Buttons/BitSplitButtonDemo.razor.css",
         "inputFile": "Pages/Components/Buttons/BitSplitButtonDemo.razor.scss"
     },
